backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.24.3
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
PyJWT==2.8.0
pytesseract==0.3.10
pydantic==2.5.0
python-jose[cryptography]==3.3.0